refactor(eve): interpret stale HITL responses centrally
